update bash scripts to run single case each time

+++ CompareWithBaseline/VerucchiOnlyRTAS2023/runVerucchiOnlyRTAS2023.sh
@@ -7,8 +7,8 @@ MinTaskNumber=3
 MaxTaskNumber=30
 TaskNumberArray=(3 4 5 6 7 8 9 10 15 20 25 30)
 ## no separator '/' at the end of the path
-ROOT_PATH="/home/zephyr/Programming/DAG_NLP" 
-# ROOT_PATH="/home/dong/workspace/DAG_NLP"
+# ROOT_PATH="/home/zephyr/Programming/DAG_NLP" 
+ROOT_PATH="/home/dong/workspace/DAG_NLP"
 # ROOT_PATH="/home/dong/workspace/DAG_batch_test/rtda_test/DAG_NLP" # for final batch test
 RESULTS_PATH="$ROOT_PATH/TaskData/dagTasks" # tBatch1's result path
 methods_dir_name=( "Initial_Res" "Verucchi_Res" ) # only do Initial and Verucchi
@@ -55,6 +55,17 @@ else
   if [[ -d scripts_and_figures_backup ]]; then rm -rf scripts_and_figures_backup; fi
   mkdir scripts_and_figures_backup
 fi
+# make directories for results
+cd $ROOT_PATH/CompareWithBaseline/$title
+for jobNumber in ${TaskNumberArray[@]}; do
+	taskset_folder_name="N$jobNumber"
+  for dir_name in ${methods_dir_name[@]}; do
+    cd $dir_name
+    if [[ ! -d $taskset_folder_name ]]; then mkdir $taskset_folder_name; fi
+    cd ..
+  done
+  if [[ ! -d dagTasks/$taskset_folder_name ]]; then mkdir dagTasks/$taskset_folder_name; fi
+done
 
 # set parameters, backup parameters and scripts parameters
 cp parameters.yaml $ROOT_PATH/sources/parameters.yaml
@@ -77,10 +88,8 @@ if [[ ! -d $ROOT_PATH/release ]]; then ./build_release_target.sh; fi
 perform_optimization() {
   # Optimize energy consumption
   cd $ROOT_PATH/release
-  cmake --build . --config Release -- -j 4
   ./tests/tBatchVerucchi # only test verucchi
   cd $ROOT_PATH/CompareWithBaseline/$title
-  sleep 1
 }
 
 # for (( jobNumber=$MinTaskNumber; jobNumber<=$MaxTaskNumber; jobNumber++ ))
@@ -91,27 +100,32 @@ do
   #   --taskSetNumber $taskSetNumber --NumberOfProcessor $coreNumberAva \
   #   --randomSeed $randomSeed --useRandomUtilization 1
 
+	echo "$title iteration is: $jobNumber"
+
+  cd $ROOT_PATH/CompareWithBaseline/$title
 	taskset_folder_name="N$jobNumber"
 
-  # instead of generate new task set, copy current tasksets into the RESULTS_PATH path
-  cd $RESULTS_PATH
-  rm *
-  cp $TASKSETS_PATH/$taskset_folder_name/* .
-  cd $ROOT_PATH/CompareWithBaseline/$title
-	
-	echo "$title iteration is: $jobNumber"
-	# initial, order optimization, verucchi
-	perform_optimization
-	
-  #copy results to corresponding folder
-  taskset_result_summary_file_name="N$jobNumber.txt"
-  for dir_name in ${methods_dir_name[@]}; do
-    cd $dir_name
-    if [[ ! -d $taskset_folder_name ]]; then mkdir $taskset_folder_name; fi
-    cp $RESULTS_PATH/*$dir_name.txt ./$taskset_folder_name/
-    rm $taskset_result_summary_file_name
-    cat ./$taskset_folder_name/* >> $taskset_result_summary_file_name
-    cd $ROOT_PATH/CompareWithBaseline/$title
+  for ((taskset_id = 0 ; taskset_id < taskSetNumber ; taskset_id++)); do
+    printf -v padded_id "%03d" $taskset_id
+
+    if [[ ! -e ./Verucchi_Res/$taskset_folder_name/dag-set-N3-$padded_id-syntheticJobs.csv_Verucchi_Res.txt ]]; then
+      # instead of generate new task set, copy current tasksets into the RESULTS_PATH path
+      cd $RESULTS_PATH
+      rm *
+      cp $TASKSETS_PATH/$taskset_folder_name/dag-set-N3-$padded_id-syntheticJobs.csv .
+      cd $ROOT_PATH/CompareWithBaseline/$title
+      perform_optimization
+      
+      #copy results to corresponding folder
+      taskset_result_summary_file_name="N$jobNumber.txt"
+      for dir_name in ${methods_dir_name[@]}; do
+        cd $dir_name
+        cp $RESULTS_PATH/dag-set-N3-$padded_id-syntheticJobs.csv_*$dir_name.txt ./$taskset_folder_name/
+        # rm $taskset_result_summary_file_name
+        cat ./$taskset_folder_name/dag-set-N3-$padded_id-syntheticJobs.csv_*$dir_name.txt >> $taskset_result_summary_file_name
+        cd $ROOT_PATH/CompareWithBaseline/$title
+      done
+    fi
   done
   if [[ ! -d dagTasks/$taskset_folder_name ]]; then mkdir dagTasks/$taskset_folder_name; fi
   cp $RESULTS_PATH/*.csv dagTasks/$taskset_folder_name/
